JCR-3313 Changed a ColumnTest test to only check for required columns
Section 6.7.1 of JSR-283 requires that "If no columns are specified, the columns available in the tabular view are implementation determined, but minimally include, for each selector, a column for each single-valued non-residual property of the selector's node type." The ColumnTest.testExpandColumnsForNodeType() method was placing additional requirements: namely that the tabular view contains **only** a column for each single-valued non-residual property of the selector's node type.

This test was changed to more correctly enforce the specification's requirements by ensuring only that there is a column for each single-valued non-residual property of the selector's node type, while allowing other columns to exist. Note that this test is slightly less strict than before, so Jackrabbit Core passes this test without any changes.

Additionally, a full build with these changes succeeds as expected.